requirements.txt
pdfplumber==0.10.2
pdf2image==1.16.3
pytesseract==0.3.10
Pillow>=9.0.0